# docs/openapi.yml
openapi: 3.0.3

info:
  title: libra
  version: 1.0.0
  description: libra API
  license:
    name: MIT

servers:
  - url: "https://digicon-libra.trap.show/"
    description: production
  - url: "http://localhost:8080/"
    description: local

paths:
  "/login":
    parameters:
      - name: username
        in: query
        required: true
        description: username
        schema:
          type: string
      - name: password
        in: query
        required: true
        description: password
        schema:
          type: string
    post:
      tags:
        - user
      summary: ノートを投稿
      responses:
        "200":
          description: ok
        "400":
          description: no
  "/whoamai":
    get:
      tags:
        - user
      summary: 私はだあれ？
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: string
                example: ramdos
  "/note":
    get:
      tags:
        - note
      summary: ノートの一覧
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/note"
    post:
      tags:
        - note
      summary: ノートを投稿
      operationId: postNote
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/postNote"
        required: true
      responses:
        "200":
          description: ok
        "405":
          description: Invalid input
  "/note/{id}":
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
          example: 13
    get:
      summary: ノートの内容
      tags:
        - note
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/noteDetail"
  "/note/author/{id}":
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
          example: 24
    get:
      tags:
        - note
      summary: 同じ著者のノート一覧
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/note"
components:
  parameters:
    id:
      name: sessions
      in: cookie
      required: true
      schema:
        type: string

  schemas:
    note:
      type: object
      properties:
        note_id:
          type: integer
        user_id:
          type: integer
        title:
          type: string
        color:
          type: string
          example: "#coffee"
        updated_at:
          type: string
          format: date
      required:
        - note_id
        - user_id
        - title
        - color
        - updated_at
    noteDetail:
      type: object
      properties:
        note_id:
          type: integer
          example: 13
        user_id:
          type: integer
          example: 24
        title:
          type: string
          example: トロッコ
        text:
          type: string
          example: 小田原熱海間に、軽便鉄道敷設の工事が始まったのは、良平の八つの年だった。良平は毎日村外れへ、その工事を見物に行った。工事を――といったところが、唯トロッコで土を運搬する――それが面白さに見に行ったのである。トロッコの上には土工が二人、土を積んだ後に佇んでいる。トロッコは山を下るのだから、人手を借りずに走って来る。煽るように車台が動いたり、土工の袢天の裾がひらついたり、細い線路がしなったり――良平はそんなけしきを眺めながら、土工になりたいと思う事がある。せめては一度でも土工と一しょに、トロッコへ乗りたいと思う事もある。トロッコは村外れの平地へ来ると、自然と其処に止まってしまう。と同時に土工たちは、身軽にトロッコを飛び降りるが早いか、その線路の終点へ車の土をぶちまける。それから今度はトロッコを押し押し、もと来た山の方へ登り始める。良平はその時乗れないまでも、押す事さえ出来たらと思うのである。或夕方、――それは二月の初旬だった。良平は二つ下の弟や、弟と同じ年の隣の子供と、トロッコの置いてある村外れへ行った。トロッコは泥だらけになったまま、薄明るい中に並んでいる。が、その外は何処を見ても、土工たちの姿は見えなかった。三人の子供は恐る恐る、一番端にあるトロッコを押した。トロッコは三人の力が揃うと、突然ごろりと車輪をまわした。良平はこの音にひやりとした。しかし二度目の車輪の音は、もう彼を驚かさなかった。ごろり、ごろり、――トロッコはそう云う音と共に、三人の手に押されながら、そろそろ線路を登って行った。その内にかれこれ十間程来ると、線路の勾配が急になり出した。トロッコも三人の力では、いくら押しても動かなくなった。どうかすれば車と一しょに、押し戻されそうにもなる事がある。良平はもう好いと思ったから、年下の二人に合図をした。「さあ、乗ろう！」彼等は一度に手をはなすと、トロッコの上へ飛び乗った。トロッコは最初徐ろに、それから見る見る勢よく、一息に線路を下り出した。その途端につき当りの風景は、忽ち両側へ分かれるように、ずんずん目の前へ展開して来る。顔に当る薄暮の風、足の下に躍るトロッコの動揺、――良平は殆ど有頂天になった。しかしトロッコは二三分の後、もうもとの終点に止まっていた。「さあ、もう一度押すじゃあ」良平は年下の二人と一しょに、又トロッコを押し上げにかかった。が、まだ車輪も動かない内に、突然彼等の後には、誰かの足音が聞え出した。のみならずそれは聞え出したと思うと、急にこう云う怒鳴り声に変った。「この野郎！　誰に断ってトロに触った？」其処には古い印袢天に、季節外れの麦藁帽をかぶった、背の高い土工が佇んでいる。――そう云う姿が目にはいった時、良平は年下の二人と一しょに、もう五六間逃げ出していた。――それぎり良平は使の帰りに、人気のない工事場のトロッコを見ても、二度と乗って見ようと思った事はない。唯その時の土工の姿は、今でも良平の頭の何処かに、はっきりした記憶を残している。薄明りの中に仄めいた、小さい黄色の麦藁帽、――しかしその記憶さえも、年毎に色彩は薄れるらしい。
        color:
          type: string
          example: "#coffee"
        updated_at:
          type: string
          format: date
          example: 2006-01-02 15:04:05
      required:
        - note_id
        - user_id
        - title
        - color
        - text
        - updated_at
    postNote:
      type: object
      properties:
        title:
          type: string
          example: トロッコ
        text:
          type: string
          example: 小田原熱海間に、軽便鉄道敷設の工事が始まったのは、良平の八つの年だった。良平は毎日村外れへ、その工事を見物に行った。工事を――といったところが、唯トロッコで土を運搬する――それが面白さに見に行ったのである。トロッコの上には土工が二人、土を積んだ後に佇んでいる。トロッコは山を下るのだから、人手を借りずに走って来る。煽るように車台が動いたり、土工の袢天の裾がひらついたり、細い線路がしなったり――良平はそんなけしきを眺めながら、土工になりたいと思う事がある。せめては一度でも土工と一しょに、トロッコへ乗りたいと思う事もある。トロッコは村外れの平地へ来ると、自然と其処に止まってしまう。と同時に土工たちは、身軽にトロッコを飛び降りるが早いか、その線路の終点へ車の土をぶちまける。それから今度はトロッコを押し押し、もと来た山の方へ登り始める。良平はその時乗れないまでも、押す事さえ出来たらと思うのである。或夕方、――それは二月の初旬だった。良平は二つ下の弟や、弟と同じ年の隣の子供と、トロッコの置いてある村外れへ行った。トロッコは泥だらけになったまま、薄明るい中に並んでいる。が、その外は何処を見ても、土工たちの姿は見えなかった。三人の子供は恐る恐る、一番端にあるトロッコを押した。トロッコは三人の力が揃うと、突然ごろりと車輪をまわした。良平はこの音にひやりとした。しかし二度目の車輪の音は、もう彼を驚かさなかった。ごろり、ごろり、――トロッコはそう云う音と共に、三人の手に押されながら、そろそろ線路を登って行った。その内にかれこれ十間程来ると、線路の勾配が急になり出した。トロッコも三人の力では、いくら押しても動かなくなった。どうかすれば車と一しょに、押し戻されそうにもなる事がある。良平はもう好いと思ったから、年下の二人に合図をした。「さあ、乗ろう！」彼等は一度に手をはなすと、トロッコの上へ飛び乗った。トロッコは最初徐ろに、それから見る見る勢よく、一息に線路を下り出した。その途端につき当りの風景は、忽ち両側へ分かれるように、ずんずん目の前へ展開して来る。顔に当る薄暮の風、足の下に躍るトロッコの動揺、――良平は殆ど有頂天になった。しかしトロッコは二三分の後、もうもとの終点に止まっていた。「さあ、もう一度押すじゃあ」良平は年下の二人と一しょに、又トロッコを押し上げにかかった。が、まだ車輪も動かない内に、突然彼等の後には、誰かの足音が聞え出した。のみならずそれは聞え出したと思うと、急にこう云う怒鳴り声に変った。「この野郎！　誰に断ってトロに触った？」其処には古い印袢天に、季節外れの麦藁帽をかぶった、背の高い土工が佇んでいる。――そう云う姿が目にはいった時、良平は年下の二人と一しょに、もう五六間逃げ出していた。――それぎり良平は使の帰りに、人気のない工事場のトロッコを見ても、二度と乗って見ようと思った事はない。唯その時の土工の姿は、今でも良平の頭の何処かに、はっきりした記憶を残している。薄明りの中に仄めいた、小さい黄色の麦藁帽、――しかしその記憶さえも、年毎に色彩は薄れるらしい。
        color:
          type: string
          example: "#coffee"
      required:
        - note_id
        - title
        - color
        - text